Fix tests, move theme to _app.tsx
