+++ pages/_app.tsx
@@ -1,11 +1,22 @@
 import '@/styles/global.css'
+import { createTheme, ThemeProvider } from '@mui/material';
 import type { AppProps } from 'next/app'
 import { RecoilRoot } from 'recoil'
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#9C426A'
+    }
+  }
+});
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
-      <Component {...pageProps} />
+      <ThemeProvider theme={theme}>
+        <Component {...pageProps} />
+      </ThemeProvider>
     </RecoilRoot>
   )
 }
